moderation-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.2.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.3
